requirements.txt
pandas>=1.5
numpy
scikit-learn
xgboost
yfinance
backtesting
alpaca-trade-api
python-dotenv
ta
